pyproject.toml
[project]
name = "fastapi-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "IngvarOrlov",email = "kampaiski@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (==2.0.39)",
    "uvicorn (==0.34.0)",
    "alembic (==1.15.2)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
